separate key classes because that's what HMX did
