+++ config/SZBE69/splits.txt
@@ -244,9 +244,33 @@ rb3/rg_messages.cpp:
 	.data       start:0x808546F0 end:0x80854A08
 	.bss        start:0x8091A254 end:0x8091A290
 
-rb3/key_messages.cpp:
-	.text       start:0x80315298 end:0x80315A70
-	.data       start:0x80854A08 end:0x80854D00
+rb3/messages/keys/keyboardkeyreleasedmsg.cpp:
+	.text       start:0x80315298 end:0x80315370
+	.data       start:0x80854CE4 end:0x80854D00
+
+rb3/messages/keys/keyboardmodmsg.cpp:
+	.text       start:0x80315370 end:0x80315448
+
+rb3/messages/keys/keyboardexpressionpedalmsg.cpp:
+	.text       start:0x80315448 end:0x80315520
+
+rb3/messages/keys/keyboardconnectedaccessories.cpp:
+	.text       start:0x80315520 end:0x803155F8
+
+rb3/messages/keys/keyboardsustainmsg.cpp:
+	.text       start:0x803155F8 end:0x803156D0
+
+rb3/messages/keys/keyboardstompbox.cpp:
+	.text       start:0x803156D0 end:0x803157A8
+
+rb3/messages/keys/keysaccelerometermsg.cpp:
+	.text       start:0x803157A8 end:0x803158C0
+
+rb3/messages/keys/keyboardlowhandplacementmsg.cpp:
+	.text       start:0x803158C0 end:0x80315998
+
+rb3/messages/keys/keyboardhighhandplacementmsg.cpp:
+	.text       start:0x80315998 end:0x80315A70
 
 rb3/dataarray.cpp:
 	.text       start:0x80315A70 end:0x803185F0
