+++ configure.py
@@ -288,7 +288,6 @@ config.libs = [
             Object(NonMatching, "rb3/formatstring.cpp"),
             Object(NonMatching, "rb3/interpolators.cpp"),
             Object(NonMatching, "rb3/joypad.cpp"),
-            Object(NonMatching, "rb3/key_messages.cpp"),
             Object(NonMatching, "rb3/netstream.cpp"),
             Object(NonMatching, "rb3/notetube.cpp"),
             Object(NonMatching, "rb3/ogg_mem.cpp"),
@@ -324,6 +323,16 @@ config.libs = [
             Object(NonMatching, "rb3/fx/fxsendpitchshift.cpp"),
             Object(NonMatching, "rb3/fx/fxsendsynapse.cpp"),
             Object(NonMatching, "rb3/fx/fxsendwah.cpp"),
+
+            Object(NonMatching, "rb3/messages/keys/keyboardkeyreleasedmsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardmodmsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardexpressionpedalmsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardconnectedaccessories.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardsustainmsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardstompbox.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keysaccelerometermsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardlowhandplacementmsg.cpp"),
+            Object(NonMatching, "rb3/messages/keys/keyboardhighhandplacementmsg.cpp"),
 
             Object(NonMatching, "rb3/unknown/800/unk_8000D980.cpp"),
             Object(NonMatching, "rb3/unknown/800/unk_8000F9B0.cpp"),
